src/main/java/ToyProject/MBTIBoard/service/MemberService.java
package ToyProject.MBTIBoard.service;

import ToyProject.MBTIBoard.domain.Member;
import ToyProject.MBTIBoard.repository.MemberRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
@Transactional
@RequiredArgsConstructor
public class MemberService {
    private final MemberRepository memberRepository;

    public Member saveMember(Member member) {
        validateDuplicateMember(member);

        return memberRepository.save(member);
    }

    private void validateDuplicateMember(Member member) {
        Member findMember = memberRepository.findByEmail(member.getEmail());
        if (findMember != null) {
            throw new IllegalStateException("이미 가입한 회원입니다");
        }
    }

    public List<Member> findAllMember() {
        return memberRepository.findAll();
    }
}
